Fix dup:check by sharing overview session helpers
