+++ src/app/overview/page.tsx
@@ -2,45 +2,12 @@ import Link from "next/link";
 import { unstable_noStore as noStore } from "next/cache";
 
 import { execFileAsync } from "@/lib/exec";
+import { inferTeamIdFromWorkspace, getActiveSessions, getAgents } from "@/lib/overview/overview-data";
 import { runOpenClaw } from "@/lib/openclaw";
 import { listRecipes } from "@/lib/recipes";
 
 export const dynamic = "force-dynamic";
 export const revalidate = 0;
-
-type AgentListItem = {
-  id: string;
-  identityName?: string | null;
-  workspace?: string | null;
-};
-
-type SessionListItem = {
-  key: string;
-  updatedAt: number;
-  ageMs?: number;
-  agentId: string;
-  kind?: string;
-  model?: string;
-  contextTokens?: number;
-  inputTokens?: number;
-  outputTokens?: number;
-  totalTokens?: number;
-};
-
-function inferTeamIdFromWorkspace(workspace: string | null | undefined) {
-  if (!workspace) return null;
-  const parts = workspace.split("/").filter(Boolean);
-  const wsPart = parts.find((p) => p.startsWith("workspace-")) ?? "";
-  if (!wsPart) return null;
-  const team = wsPart.slice("workspace-".length);
-  return team || null;
-}
-
-async function getAgents(): Promise<AgentListItem[]> {
-  const res = await runOpenClaw(["agents", "list", "--json"]);
-  if (!res.ok) return [];
-  return JSON.parse(res.stdout) as AgentListItem[];
-}
 
 async function getTeamsFromRecipes(): Promise<{ teamNames: Record<string, string> }> {
   const items = await listRecipes();
@@ -52,13 +19,6 @@ async function getTeamsFromRecipes(): Promise<{ teamNames: Record<string, string
     teamNames[r.id] = name;
   }
   return { teamNames };
-}
-
-async function getActiveSessions(minutes: number): Promise<SessionListItem[]> {
-  const res = await runOpenClaw(["sessions", "--active", String(minutes), "--all-agents", "--json"]);
-  if (!res.ok) return [];
-  const parsed = JSON.parse(res.stdout) as { sessions?: SessionListItem[] };
-  return Array.isArray(parsed.sessions) ? parsed.sessions : [];
 }
 
 type GatewayStatusSummary = {
